requirements.txt
jupyterlab>=3.4.5
pandas>=1.2.4
pyyaml>=5.4.1
argparse>=1.4.0
web3>=5.19.0
requests>=2.25.1
networkx>=2.5.1
scipy>=1.7.0
numpy>=1.20.3
powerlaw>=1.4.6
joblib>=1.0.1
scikit-learn>=0.24.2
seaborn>=0.11.2
python-igraph>=0.9.6
pydot>=1.4.1